Improve mechanism to send and receive a packet of variables
Change the mechanism that triggers redraw - ensuring that all data items are received instead of checking for shape of main data item. Also improvement in checking for lost data, and automatic reload of figure if this happens. A bug fix related to sendfigure requests - operates on a per figure basis instead of shared flag for all figures.