Figure width glitch workaround
A workaround for a glitch that occurs when loading figures. It is difficult to know what the axiscanvas.width is at the time when applyviewlayout is executed because the figure show title, label etc information is not yet available. Therefore instead use figure width which is independent of these view settings. This means that a bit more x-axis data samples than necessary is sent over. A glitch occurs when it is detected that the figure width changes (performs reload) from initial applyviewlayout update to subsequent update. For slow internet connections this is noticeable, because effectively the page data is loaded twice instead of once (baseline matrix pages).